requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.13.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.7
click==8.3.1
colorama==0.4.6
cryptography==46.0.6
distlib==0.4.0
distro==1.9.0
einops==0.8.2
fastapi==0.135.3
filelock==3.25.2
filetype==1.2.0
fs_extra==0.0.4
fsspec==2026.3.0
ftfy==6.3.1
google-auth==2.49.1
google-genai==1.70.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
identify==2.6.18
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
markdown2==2.5.5
markdownify==1.2.2
marker-pdf==1.10.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
nodeenv==1.10.0
numpy==2.4.4
openai==1.109.1
opencv-python-headless==4.11.0.86
packaging==26.0
pdftext==0.6.3
pillow==10.4.0
platformdirs==4.9.4
pre_commit==4.5.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pypdfium2==4.30.0
python-discovery==1.2.1
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2024.11.6
requests==2.33.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==81.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
starlette==1.0.0
surya-ocr==0.17.1
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
virtualenv==21.2.0
wcwidth==0.6.0
websockets==16.0
